better logic

+++ gentoo/oracle-cloud/world-update.sh
@@ -30,8 +30,8 @@ zcat /proc/config.gz > /usr/src/linux/.config
 make olddefconfig && make -j $cpu_cores
 grep -q "BOOT_IMAGE=[^ ]*[.]prev" /proc/cmdline || {
     rm -f /boot/*.prev
-    files=$(ls /boot/*-$(uname -r)) && for file in $files; do
-        cp $file ${file}.prev
+    for file in /boot/*-$(uname -r); do
+        [ -f "$file" ] && cp $file ${file}.prev
     done
 }
 make install
@@ -41,7 +41,7 @@ printf "%s\n" > /boot/grub/grub.cfg.new \
     "terminal_input serial" \
     "terminal_output serial" \
     "set timeout=5"
-files=$(ls -t /boot/vmlinuz-*|grep -v "$(uname -r)$") && for file in $files; do
+for file in $(ls -t /boot/vmlinuz-*|grep -v "$(uname -r)$"); do
     printf "%s\n" \
         "menuentry '$file' {" \
         "  linux $file root=/dev/sda1 console=tty1 console=ttyS0 nvme.shutdown_timeout=10 libiscsi.debug_libiscsi_eh=1" \
